+++ lib/eventum/class.issue.php
@@ -1662,6 +1662,12 @@ class Issue
         // clear project cache
         self::getProjectID($issue_id, true);
 
+        History::add($issue_id, Auth::getUserID(), 'issue_moved', 'Issue moved from {old_project} to {new_project} by {user}', [
+                'old_project' => Project::getName($current_prj_id),
+                'new_project' => Project::getName($new_prj_id),
+                'user' => User::getFullName(Auth::getUserID()),
+        ]);
+
         Workflow::handleIssueMovedFromProject($current_prj_id, $issue_id, $new_prj_id);
         Workflow::handleIssueMovedToProject($new_prj_id, $issue_id, $current_prj_id);